Fixed wrong cast

+++ glscopeclient/ScopeSyncWizard.cpp
@@ -337,7 +337,7 @@ void ScopeSyncWizard::OnWaveformDataReady()
 	*/
 	m_maxSkewSamples = static_cast<int64_t>(pw->m_offsets.size() / 2);
 	
-	m_maxSkewSamples = min(m_maxSkewSamples, 10000LL);
+	m_maxSkewSamples = min(m_maxSkewSamples, static_cast<int64_t>(10000LL));
 	
 	m_delta = - m_maxSkewSamples;
